display: Fix scale by rounding (Fixes #258)

+++ common/xfce-randr.h
@@ -68,8 +68,8 @@ struct _XfceRandr
 
     /* selected settings for all connected outputs */
     RRMode              *mode;
-    gfloat              *scalex;
-    gfloat              *scaley;
+    gdouble             *scalex;
+    gdouble             *scaley;
     Rotation            *rotation;
     Rotation            *rotations;
     XfceOutputPosition  *position;
